Fix wget command not found error
Added `wget` to the dependencies list to resolve the following error on fresh default installations of Debian 13.
```
/usr/lib/debian-nvidia-installer/cudarepo.sh: line 95: wget: command not found
CRITICAL: Failed to download the CUDA repository installation package. Operation aborted.
```

The error occurs on fresh default installations of Debian 13 while installing any of the 575 or 580 driver flavors because the `usr/lib/debian-nvidia-installer/cudarepo.sh` script uses `wget` to fetch the keyring but `wget` is not installed by default.